Added exercise chaining, removed category exercise
Includes many more smaller changes such as removing the option to choose how many words are in an exercise, and allowing what language the answers are in for the audio exercise to be chosen. I also realised that I've been spelling exercise as "excercise", but I should have corrected all of the misspellings.